Updated to V3
- Added presets for Navigation, Apps, Power Control, Pairing, and Volume Control.
- Added an action to set the Volume Level specifically.
- Added a feedback to watch the Volume Level.
