+++ .gitignore
@@ -1,2 +1,4 @@
 node_modules/
 package-lock.json
+/pkg
+/pkg.tgz
