+++ upgrades.js
@@ -1,0 +1,13 @@
+module.exports = [
+	/*
+	 * Place your upgrade scripts here
+	 * Remember that once it has been added it cannot be removed!
+	 */
+	// function (context, props) {
+	// 	return {
+	// 		updatedConfig: null,
+	// 		updatedActions: [],
+	// 		updatedFeedbacks: [],
+	// 	}
+	// },
+]
